+++ streamloots/dependency-reduced-pom.xml
@@ -1,10 +1,14 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+  <parent>
+    <artifactId>streamlootsplugins</artifactId>
+    <groupId>fr.entrivax</groupId>
+    <version>0.0.1-SNAPSHOT</version>
+  </parent>
   <modelVersion>4.0.0</modelVersion>
   <groupId>fr.entrivax</groupId>
   <artifactId>streamloots</artifactId>
   <name>Streamloots</name>
-  <version>0.0.1-SNAPSHOT</version>
   <build>
     <plugins>
       <plugin>
