+++ streamloots/pom.xml
@@ -1,0 +1,85 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+	<groupId>fr.entrivax</groupId>
+	<artifactId>streamloots</artifactId>
+	<name>Streamloots</name>
+	<packaging>jar</packaging>
+
+	<parent>
+		<groupId>fr.entrivax</groupId>
+		<artifactId>streamlootsplugins</artifactId>
+		<version>0.0.1-SNAPSHOT</version>
+	</parent>
+
+	<repositories>
+		<repository>
+			<id>papermc</id>
+			<url>https://papermc.io/repo/repository/maven-public/</url>
+		</repository>
+	</repositories>
+	<dependencies>
+		<dependency>
+			<groupId>io.papermc.paper</groupId>
+			<artifactId>paper-api</artifactId>
+			<version>1.17.1-R0.1-SNAPSHOT</version>
+			<scope>provided</scope>
+		</dependency>
+		<dependency>
+		    <groupId>org.apache.httpcomponents</groupId>
+		    <artifactId>httpclient</artifactId>
+		    <version>4.5.13</version>
+		</dependency>
+		<dependency>
+		    <groupId>commons-io</groupId>
+		    <artifactId>commons-io</artifactId>
+		    <version>2.8.0</version>
+		</dependency>
+		<dependency>
+			<groupId>com.google.code.gson</groupId>
+			<artifactId>gson</artifactId>
+			<version>2.8.7</version>
+		</dependency>
+	</dependencies>
+	<pluginRepositories>
+	    <pluginRepository>
+	        <id>maven-snapshots</id>
+	        <url>https://repository.apache.org/content/repositories/snapshots/</url>
+	    </pluginRepository>
+	</pluginRepositories>
+	<build>
+		<plugins>
+			<plugin>
+			    <groupId>org.apache.maven.plugins</groupId>
+			    <artifactId>maven-shade-plugin</artifactId>
+			    <version>3.3.0-SNAPSHOT</version>
+			    <executions>
+			        <execution>
+			            <phase>package</phase>
+			            <goals>
+			                <goal>shade</goal>
+			            </goals>
+			            <configuration>
+			                <finalName>Streamloots</finalName>
+			                <relocations>
+			                    <relocation>
+			                        <pattern>org.apache.commons</pattern>
+			                        <shadedPattern>me.org.apache.commons</shadedPattern>
+			                    </relocation>
+			                    <relocation>
+			                        <pattern>org.apache.http</pattern>
+			                        <shadedPattern>me.org.apache.http</shadedPattern>
+			                    </relocation>
+			                </relocations>
+			            </configuration>
+			        </execution>
+			    </executions>
+			</plugin>
+		</plugins>
+	</build>
+	<properties>
+		<maven.compiler.source>1.8</maven.compiler.source>
+		<maven.compiler.target>1.8</maven.compiler.target>
+	</properties>
+</project>
