+++ streamloots/src/main/java/fr/entrivax/streamloots/CardCommand.java
