+++ streamloots/src/main/java/fr/entrivax/streamloots/CardConfig.java
