+++ streamloots/src/main/java/fr/entrivax/streamloots/EntryPoint.java
