+++ streamloots/src/main/java/fr/entrivax/streamloots/IStreamlootsRedemptionHandler.java
