+++ streamloots/src/main/java/fr/entrivax/streamloots/StreamlootsClient.java
