+++ streamloots/src/main/java/fr/entrivax/streamloots/StreamlootsClientManager.java
