+++ streamloots/src/main/java/fr/entrivax/streamloots/StreamlootsRedemptionHandler.java
