+++ streamloots/src/main/java/fr/entrivax/streamloots/commands/IStreamlootsCardCommand.java
