+++ streamloots/src/main/java/fr/entrivax/streamloots/commands/StreamlootsCardSetHungerCommand.java
