+++ streamloots/src/main/resources/cards.schema.json
