+++ streamloots/src/main/resources/plugin.yml
