+++ streamlootsbase/src/main/resources/plugin.yml
@@ -1,0 +1,6 @@
+main: "fr.entrivax.streamlootsbase.StreamlootsBasePlugin"
+author: Entrivax
+version: 0.1
+name: "Streamloots-Base"
+depend: ["Streamloots"]
+api-version: 1.17
